Implement naive frequency-based agent
This agent plays (uniformly) randomly for the first 50 steps.
After that it uses the relative frequencies of the opponent's previous actions as a probability distribution to chose it's own actions from. If this approach doesn't work well, we start playing randomly again.
The decision whether we use frequencies or not is flipped with a certain probability.